Added Dummy API and run SpringBoot as a server
