+++ backendserver/README.md
@@ -1,0 +1,3 @@
+To run:
+
+```./mvnw clean spring-boot:run```
