+++ backendserver/src/main/java/mariomonday/backend/database/schema/GameType.java
